+++ crates/domain_manager/src/utils/mod.rs
@@ -1,3 +1,4 @@
+pub mod aliyun_utils;
 pub mod error_logger;
 pub mod formatted_strings;
 pub mod i18_utils;